add interactive.js
In this interactive file, we give the user 3 parameters that are coefficients to solve a quadratic equation. getParams function collects the input.
validateInput checks the correctness of the parameters and if something is wrong, displays an error and uses getParams again to re-enter the parameters and
returns them as an object. evalEquation contains the key logic for solving the equation and outputs the answer to the console.